requirements.txt
streamlit
ibm_db
bcrypt
python-dotenv
pandas
plotly
ibm-watsonx-ai
requests
numpy<2
